+++ DbSqlHelper/DbSqlHelperSetting.cs
@@ -1,0 +1,7 @@
+﻿namespace DbSqlHelper
+{
+    public static class DbSqlHelperSetting
+    {
+        public static int DefaultCommandTimeout = 60;
+    }
+}