+++ Changelog.md
@@ -1,3 +1,8 @@
+#### Version 0.9.4
+- [X] Add ParameterExtension Dapper Style AddParams
+- [X] Support AddConnection(Type connectionType,string connectionString) 
+- [X] Add Editable DefaultCommandTimeout
+
 #### Version 0.9.3
 - [X] Add String Extension For Easy Execute Sql